requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-decouple==3.8

